refactor: move getReadingTime alongside the place where it's being used
